- fix various compilation warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@52108 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/cleanup/cleanup_utils.hpp
@@ -231,7 +231,7 @@ bool RemoveDupsNoSort(Cont& l, bool case_insensitive = false)
 
 #define COMPRESS_STRING_MEMBER(o, x) \
     if ((o).IsSet##x()) { \
-        const int old_len = (o).Set##x().length(); \
+        const size_t old_len = (o).Set##x().length(); \
         CompressSpaces((o).Set##x()); \
         if( old_len != (o).Set##x().length() ) { \
             ChangeMade(CCleanupChange::eTrimSpaces); \
